Move svg components into the timeline as well
